+++ app/app/src/main/AndroidManifest.xml
@@ -34,6 +34,11 @@
             android:launchMode="singleTop"
             android:windowSoftInputMode="adjustPan" />
         <activity
+            android:name="com.surverior.android.activity.ProfileActivity"
+            android:label="@string/app_name"
+            android:launchMode="singleTop"
+            android:windowSoftInputMode="adjustPan" />
+        <activity
             android:name="com.surverior.android.activity.MainActivity"
             android:label="@string/app_name"
             android:launchMode="singleTop" />
