+++ app/app/src/main/java/com/surverior/android/app/AppConfig.java
@@ -2,8 +2,11 @@ package com.surverior.android.app;
 
 public class AppConfig {
 	// Server user login url
-	public static String URL_LOGIN = "http://192.168.173.1/android_login_api/login.php";
+	public static String URL_LOGIN = "http://192.168.100.2/android_login_api/login.php";
 
 	// Server user register url
-	public static String URL_REGISTER = "http://192.168.173.1/android_login_api/register.php";
+	public static String URL_REGISTER = "http://192.168.100.2/android_login_api/register.php";
+
+	//Server user update url
+	public static String URL_UPDATE = "http://192.168.100.2/android_login_api/profile.php";
 }
