+++ app/app/src/main/java/com/surverior/android/helper/SQLiteHandler.java
@@ -30,15 +30,15 @@ public class SQLiteHandler extends SQLiteOpenHelper {
 
 	// Login Table Columns names
 	private static final String KEY_ID = "id";
-	private static final String KEY_NAME = "name";
+//	private static final String KEY_NAME = "name";
 	private static final String KEY_EMAIL = "email";
 	private static final String KEY_UID = "uid";
-	private static final String KEY_GENDER = "gender";
-	private static final String KEY_BIRTH_DATE = "birth_date";
-	private static final String KEY_PROFESSION = "profession";
-	private static final String KEY_CITY = "city";
-	private static final String KEY_PROVINCE = "province";
-	private static final String KEY_CREATED_AT = "created_at";
+//	private static final String KEY_GENDER = "gender";
+//	private static final String KEY_BIRTH_DATE = "birth_date";
+//	private static final String KEY_PROFESSION = "profession";
+//	private static final String KEY_CITY = "city";
+//	private static final String KEY_PROVINCE = "province";
+//	private static final String KEY_CREATED_AT = "created_at";
 
 	public SQLiteHandler(Context context) {
 		super(context, DATABASE_NAME, null, DATABASE_VERSION);
@@ -48,12 +48,8 @@ public class SQLiteHandler extends SQLiteOpenHelper {
 	@Override
 	public void onCreate(SQLiteDatabase db) {
 		String CREATE_LOGIN_TABLE = "CREATE TABLE " + TABLE_USER + "("
-				+ KEY_ID + " INTEGER PRIMARY KEY," + KEY_NAME + " TEXT,"
-				+ KEY_EMAIL + " TEXT UNIQUE," + KEY_UID + " TEXT,"
-				+ KEY_GENDER + " TEXT," + KEY_BIRTH_DATE + " TEXT,"
-				+ KEY_PROFESSION + " TEXT," + KEY_CITY + " TEXT,"
-				+ KEY_PROVINCE + " TEXT,"
-				+ KEY_CREATED_AT + " TEXT" + ")";
+				+ KEY_ID + " INTEGER PRIMARY KEY,"
+				+ KEY_EMAIL + " TEXT UNIQUE," + KEY_UID + " TEXT" + ")";
 		db.execSQL(CREATE_LOGIN_TABLE);
 
 		Log.d(TAG, "Database tables created");
@@ -72,13 +68,13 @@ public class SQLiteHandler extends SQLiteOpenHelper {
 	/**
 	 * Storing user details in database
 	 * */
-	public void addUser(String name, String email, String uid/*,
+	public void addUser(String email, String uid/*,
 						String gender,String birth_date,String profession,
 						String city, String province,  String created_at*/) {
 		SQLiteDatabase db = this.getWritableDatabase();
 
 		ContentValues values = new ContentValues();
-		values.put(KEY_NAME, name); // Name
+		//values.put(KEY_NAME, name); // Name
 		values.put(KEY_EMAIL, email); // Email
 		values.put(KEY_UID, uid); // Email/*
 		/*values.put(KEY_GENDER, gender); // Gender
@@ -107,7 +103,7 @@ public class SQLiteHandler extends SQLiteOpenHelper {
 		// Move to first row
 		cursor.moveToFirst();
 		if (cursor.getCount() > 0) {
-			user.put("name", cursor.getString(1));
+			//user.put("name", cursor.getString(1));
 			user.put("email", cursor.getString(2));
 			user.put("uid", cursor.getString(3));
 			user.put("created_at", cursor.getString(4));
