+++ app/app/src/main/res/layout/activity_register.xml
@@ -28,18 +28,6 @@
             android:paddingLeft="20dp"
             android:paddingRight="20dp" >
 
-            <EditText
-                android:id="@+id/name"
-                android:layout_width="fill_parent"
-                android:layout_height="wrap_content"
-                android:layout_marginBottom="10dp"
-                android:background="@color/input_register_bg"
-                android:hint="@string/hint_name"
-                android:padding="10dp"
-                android:singleLine="true"
-                android:inputType="textCapWords"
-                android:textColor="@color/input_register"
-                android:textColorHint="@color/input_register_hint" />
 
             <EditText
                 android:id="@+id/email"
@@ -81,103 +69,7 @@
                 android:textColor="@color/input_register"
                 android:textColorHint="@color/input_register_hint" />
 
-            <TextView
-            android:layout_width="fill_parent"
-            android:layout_height="wrap_content"
-            android:text="Gender"
-            android:textColor="@color/input_register"
-            android:textAppearance="?android:attr/textAppearanceMedium" />
 
-            <RadioGroup
-                android:id="@+id/radioGroup"
-                android:layout_width="fill_parent"
-                android:layout_height="wrap_content"
-                android:orientation="vertical">
-
-                <RadioButton
-                    android:id="@+id/radioButton1"
-                    android:layout_width="fill_parent"
-                    android:layout_height="wrap_content"
-                    android:textColor="@color/input_register"
-                    android:text="Male" />
-
-                <RadioButton
-                    android:id="@+id/radioButton2"
-                    android:layout_width="fill_parent"
-                    android:layout_height="wrap_content"
-                    android:textColor="@color/input_register"
-                    android:text="Female" />
-            </RadioGroup>
-
-            <LinearLayout
-                android:layout_width="fill_parent"
-                android:layout_height="wrap_content"
-                android:orientation="horizontal" >
-
-                <EditText
-                    android:id="@+id/birthdate"
-                    android:layout_width="250dp"
-                    android:layout_height="wrap_content"
-                    android:layout_marginBottom="10dp"
-                    android:background="@color/input_register_bg"
-                    android:text="@string/hint_birthdate"
-                    android:layout_gravity="left"
-                    android:padding="10dp"
-                    android:textColor="@color/input_register"
-                    android:editable="false"
-                    android:textColorHint="@color/input_register_hint" />
-                <ImageButton
-                    android:id="@+id/imageButton1"
-                    android:layout_width="fill_parent"
-                    android:layout_height="wrap_content"
-                    android:layout_gravity="right"
-                    android:contentDescription="@string/selectdate"
-                    android:cropToPadding="true"
-                    android:src="@drawable/calendar_icon3"/>
-
-            </LinearLayout>
-
-            <Spinner
-                android:id="@+id/job"
-                android:layout_width="fill_parent"
-                android:layout_height="wrap_content"
-                android:background="@color/input_register_bg"
-                android:hint="@string/hint_job"
-                android:padding="10dp"
-                android:layout_marginBottom="10dp"
-                android:singleLine="true"
-                android:spinnerMode="dialog"
-                android:prompt="@string/hint_job"
-                android:text="@string/hint_job"
-                android:textColor="@color/input_register"
-                />
-
-            <EditText
-                android:id="@+id/city"
-                android:layout_width="fill_parent"
-                android:layout_height="wrap_content"
-                android:layout_marginBottom="10dp"
-                android:background="@color/input_register_bg"
-                android:hint="@string/hint_city"
-                android:inputType="textCapWords"
-                android:padding="10dp"
-                android:singleLine="true"
-                android:textColor="@color/input_register"
-                android:textColorHint="@color/input_register_hint" />
-
-            <Spinner
-                android:id="@+id/province"
-                android:layout_width="fill_parent"
-                android:layout_height="wrap_content"
-                android:background="@color/input_register_bg"
-                android:hint="@string/hint_province"
-                android:padding="10dp"
-                android:singleLine="true"
-                android:spinnerMode="dialog"
-                android:prompt="@string/hint_province"
-                android:text="@string/hint_province"
-                android:textColor="@color/input_register"
-                />
